+++ webapi/src/main/scala/org/knora/webapi/responders/v2/OntologyResponderV2.scala
@@ -41,6 +41,9 @@ import util.cache.CacheUtil
 
 import akka.http.scaladsl.util.FastFuture
 import akka.pattern._
+import org.knora.webapi.responders.v2.ontology.Cache.{ONTOLOGY_CACHE_LOCK_IRI, OntologyCacheData, makeOntologyCache}
+import org.knora.webapi.responders.v2.ontology.{Cache, DeleteCardinalitiesFromClass}
+import org.knora.webapi.responders.v2.ontology.Ontology.OntologyGraph
 
 import java.time.Instant
 import scala.collection.immutable
@@ -67,46 +70,12 @@ import scala.concurrent.Future
   */
 class OntologyResponderV2(responderData: ResponderData) extends Responder(responderData) {
 
-  // The name of the ontology cache.
-  private val OntologyCacheName = "ontologyCache"
-
-  // The cache key under which cached ontology data is stored.
-  private val OntologyCacheKey = "ontologyCacheData"
-
-  // The global ontology cache lock. This is needed because every ontology update replaces the whole ontology cache
-  // (because definitions in one ontology can refer to definitions in another ontology). Without a global lock,
-  // concurrent updates (even to different ontologies) could overwrite each other.
-  private val ONTOLOGY_CACHE_LOCK_IRI = "http://rdfh.ch/ontologies"
-
-  /**
-    * The in-memory cache of ontologies.
-    *
-    * @param ontologies              a map of ontology IRIs to ontologies.
-    * @param subClassOfRelations     a map of subclasses to their base classes.
-    * @param superClassOfRelations   a map of base classes to their subclasses.
-    * @param subPropertyOfRelations  a map of subproperties to their base proeprties.
-    * @param guiAttributeDefinitions a map of salsah-gui:Guielement individuals to their GUI attribute definitions.
-    * @param standoffProperties      a set of standoff properties.
-    */
-  private case class OntologyCacheData(ontologies: Map[SmartIri, ReadOntologyV2],
-                                       subClassOfRelations: Map[SmartIri, Seq[SmartIri]],
-                                       superClassOfRelations: Map[SmartIri, Set[SmartIri]],
-                                       subPropertyOfRelations: Map[SmartIri, Set[SmartIri]],
-                                       guiAttributeDefinitions: Map[SmartIri, Set[SalsahGuiAttributeDefinition]],
-                                       standoffProperties: Set[SmartIri]) {
-    lazy val allPropertyDefs: Map[SmartIri, PropertyInfoContentV2] = ontologies.values
-      .flatMap(_.properties.map {
-        case (propertyIri, readPropertyInfo) => propertyIri -> readPropertyInfo.entityInfoContent
-      })
-      .toMap
-  }
-
   /**
     * Receives a message of type [[OntologiesResponderRequestV2]], and returns an appropriate response message.
     */
   def receive(msg: OntologiesResponderRequestV2) = msg match {
     case LoadOntologiesRequestV2(featureFactoryConfig, requestingUser) =>
-      loadOntologies(featureFactoryConfig, requestingUser)
+      Cache.loadOntologies(settings, storeManager, featureFactoryConfig, requestingUser)
     case EntityInfoGetRequestV2(classIris, propertyIris, requestingUser) =>
       getEntityInfoResponseV2(classIris, propertyIris, requestingUser)
     case StandoffEntityInfoGetRequestV2(standoffClassIris, standoffPropertyIris, requestingUser) =>
@@ -156,1133 +125,6 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
     case canDeleteOntologyRequest: CanDeleteOntologyRequestV2 => canDeleteOntology(canDeleteOntologyRequest)
     case deleteOntologyRequest: DeleteOntologyRequestV2       => deleteOntology(deleteOntologyRequest)
     case other                                                => handleUnexpectedMessage(other, log, this.getClass.getName)
-  }
-
-  /**
-    * Represents the contents of a named graph representing an ontology.
-    *
-    * @param ontologyIri       the ontology IRI, which is also the IRI of the named graph.
-    * @param constructResponse the triplestore's response to a CONSTRUCT query that gets the contents of the named graph.
-    */
-  private case class OntologyGraph(ontologyIri: SmartIri, constructResponse: SparqlExtendedConstructResponse)
-
-  /**
-    * Loads and caches all ontology information.
-    *
-    * @param featureFactoryConfig the feature factory configuration.
-    * @param requestingUser       the user making the request.
-    * @return a [[SuccessResponseV2]].
-    */
-  private def loadOntologies(featureFactoryConfig: FeatureFactoryConfig,
-                             requestingUser: UserADM): Future[SuccessResponseV2] = {
-    val loadOntologiesFuture: Future[SuccessResponseV2] = for {
-      _ <- Future {
-        if (!(requestingUser.id == KnoraSystemInstances.Users.SystemUser.id || requestingUser.permissions.isSystemAdmin)) {
-          throw ForbiddenException(s"Only a system administrator can reload ontologies")
-        }
-      }
-
-      // Get all ontology metadata.
-      allOntologyMetadataSparql <- FastFuture.successful(
-        org.knora.webapi.messages.twirl.queries.sparql.v2.txt
-          .getAllOntologyMetadata(triplestore = settings.triplestoreType)
-          .toString())
-      allOntologyMetadataResponse: SparqlSelectResult <- (storeManager ? SparqlSelectRequest(allOntologyMetadataSparql))
-        .mapTo[SparqlSelectResult]
-      allOntologyMetadata: Map[SmartIri, OntologyMetadataV2] = buildOntologyMetadata(allOntologyMetadataResponse)
-
-      knoraBaseOntologyMetadata: OntologyMetadataV2 = allOntologyMetadata.getOrElse(
-        OntologyConstants.KnoraBase.KnoraBaseOntologyIri.toSmartIri,
-        throw InconsistentRepositoryDataException(s"No knora-base ontology found"))
-      knoraBaseOntologyVersion: String = knoraBaseOntologyMetadata.ontologyVersion.getOrElse(
-        throw InconsistentRepositoryDataException(
-          "The knora-base ontology in the repository is not up to date. See the Knora documentation on repository updates."))
-
-      _ = if (knoraBaseOntologyVersion != KnoraBaseVersion) {
-        throw InconsistentRepositoryDataException(
-          s"The knora-base ontology in the repository has version '$knoraBaseOntologyVersion', but this version of Knora requires '$KnoraBaseVersion'. See the Knora documentation on repository updates.")
-      }
-
-      // Get the contents of each named graph containing an ontology.
-      ontologyGraphResponseFutures: Iterable[Future[OntologyGraph]] = allOntologyMetadata.keys.map { ontologyIri =>
-        val ontologyGraphConstructQuery = org.knora.webapi.messages.twirl.queries.sparql.v2.txt
-          .getOntologyGraph(
-            triplestore = settings.triplestoreType,
-            ontologyGraph = ontologyIri
-          )
-          .toString
-
-        (storeManager ? SparqlExtendedConstructRequest(
-          sparql = ontologyGraphConstructQuery,
-          featureFactoryConfig = featureFactoryConfig
-        )).mapTo[SparqlExtendedConstructResponse].map { response =>
-          OntologyGraph(ontologyIri = ontologyIri, constructResponse = response)
-        }
-      }
-
-      ontologyGraphs: Iterable[OntologyGraph] <- Future.sequence(ontologyGraphResponseFutures)
-
-      _ = makeOntologyCache(allOntologyMetadata, ontologyGraphs)
-    } yield SuccessResponseV2("Ontologies loaded.")
-
-    loadOntologiesFuture.recover {
-      case exception: Throwable =>
-        exception match {
-          case inconsistentRepositoryDataException: InconsistentRepositoryDataException =>
-            log.error(inconsistentRepositoryDataException.message)
-            SuccessResponseV2(
-              s"An error occurred when loading ontologies: ${inconsistentRepositoryDataException.message}")
-
-          case other => throw other
-        }
-    }
-  }
-
-  /**
-    * Given ontology metdata and ontology graphs read from the triplestore, constructs the ontology cache.
-    *
-    * @param allOntologyMetadata a map of ontology IRIs to ontology metadata.
-    * @param ontologyGraphs      a list of ontology graphs.
-    */
-  private def makeOntologyCache(allOntologyMetadata: Map[SmartIri, OntologyMetadataV2],
-                                ontologyGraphs: Iterable[OntologyGraph]): Unit = {
-    // Get the IRIs of all the entities in each ontology.
-
-    // A map of ontology IRIs to class IRIs in each ontology.
-    val classIrisPerOntology: Map[SmartIri, Set[SmartIri]] = getEntityIrisFromOntologyGraphs(
-      ontologyGraphs = ontologyGraphs,
-      entityTypes = Set(OntologyConstants.Owl.Class)
-    )
-
-    // A map of ontology IRIs to property IRIs in each ontology.
-    val propertyIrisPerOntology: Map[SmartIri, Set[SmartIri]] = getEntityIrisFromOntologyGraphs(
-      ontologyGraphs = ontologyGraphs,
-      entityTypes = Set(
-        OntologyConstants.Owl.ObjectProperty,
-        OntologyConstants.Owl.DatatypeProperty,
-        OntologyConstants.Owl.AnnotationProperty,
-        OntologyConstants.Rdf.Property
-      )
-    )
-
-    // A map of ontology IRIs to named individual IRIs in each ontology.
-    val individualIrisPerOntology: Map[SmartIri, Set[SmartIri]] = getEntityIrisFromOntologyGraphs(
-      ontologyGraphs = ontologyGraphs,
-      entityTypes = Set(OntologyConstants.Owl.NamedIndividual)
-    )
-
-    // Construct entity definitions.
-
-    // A map of class IRIs to class definitions.
-    val allClassDefs: Map[SmartIri, ClassInfoContentV2] = ontologyGraphs.flatMap { ontologyGraph =>
-      constructResponseToClassDefinitions(
-        classIris = classIrisPerOntology(ontologyGraph.ontologyIri),
-        constructResponse = ontologyGraph.constructResponse
-      )
-    }.toMap
-
-    // A map of property IRIs to property definitions.
-    val allPropertyDefs: Map[SmartIri, PropertyInfoContentV2] = ontologyGraphs.flatMap { ontologyGraph =>
-      constructResponseToPropertyDefinitions(
-        propertyIris = propertyIrisPerOntology(ontologyGraph.ontologyIri),
-        constructResponse = ontologyGraph.constructResponse
-      )
-    }.toMap
-
-    // A map of OWL named individual IRIs to named individuals.
-    val allIndividuals: Map[SmartIri, IndividualInfoContentV2] = ontologyGraphs.flatMap { ontologyGraph =>
-      constructResponseToIndividuals(
-        individualIris = individualIrisPerOntology(ontologyGraph.ontologyIri),
-        constructResponse = ontologyGraph.constructResponse
-      )
-    }.toMap
-
-    // A map of salsah-gui:Guielement individuals to their GUI attribute definitions.
-    val allGuiAttributeDefinitions: Map[SmartIri, Set[SalsahGuiAttributeDefinition]] = makeGuiAttributeDefinitions(
-      allIndividuals)
-
-    // Determine relations between entities.
-
-    // A map of class IRIs to their immediate base classes.
-    val directSubClassOfRelations: Map[SmartIri, Set[SmartIri]] = allClassDefs.map {
-      case (classIri, classDef) => classIri -> classDef.subClassOf
-    }
-
-    // A map of property IRIs to their immediate base properties.
-    val directSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]] = allPropertyDefs.map {
-      case (propertyIri, propertyDef) => propertyIri -> propertyDef.subPropertyOf
-    }
-
-    val allClassIris = allClassDefs.keySet
-    val allPropertyIris = allPropertyDefs.keySet
-
-    // A map in which each class IRI points to the full sequence of its base classes.
-    val allSubClassOfRelations: Map[SmartIri, Seq[SmartIri]] = allClassIris.toSeq.map { classIri =>
-      // get the hierarchically ordered base classes.
-      val baseClasses: Seq[SmartIri] = OntologyUtil.getAllBaseDefs(classIri, directSubClassOfRelations)
-      // prepend the classIri to the sequence of base classes because a class is also a subclass of itself.
-      (classIri, classIri +: baseClasses)
-    }.toMap
-
-    // A map in which each class IRI points to the full set of its subclasses. A class is also
-    // a subclass of itself.
-    val allSuperClassOfRelations: Map[SmartIri, Set[SmartIri]] = calculateSuperClassOfRelations(allSubClassOfRelations)
-
-    // Make a map in which each property IRI points to the full set of its base properties. A property is also
-    // a subproperty of itself.
-    val allSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]] = allPropertyIris.map { propertyIri =>
-      (propertyIri, OntologyUtil.getAllBaseDefs(propertyIri, directSubPropertyOfRelations).toSet + propertyIri)
-    }.toMap
-
-    // A set of all subproperties of knora-base:resourceProperty.
-    val allKnoraResourceProps: Set[SmartIri] = allPropertyIris.filter { prop =>
-      val allPropSubPropertyOfRelations = allSubPropertyOfRelations(prop)
-      prop == OntologyConstants.KnoraBase.ResourceProperty.toSmartIri ||
-      allPropSubPropertyOfRelations.contains(OntologyConstants.KnoraBase.HasValue.toSmartIri) ||
-      allPropSubPropertyOfRelations.contains(OntologyConstants.KnoraBase.HasLinkTo.toSmartIri)
-    }
-
-    // A set of all subproperties of knora-base:hasLinkTo.
-    val allLinkProps: Set[SmartIri] = allPropertyIris.filter(prop =>
-      allSubPropertyOfRelations(prop).contains(OntologyConstants.KnoraBase.HasLinkTo.toSmartIri))
-
-    // A set of all subproperties of knora-base:hasLinkToValue.
-    val allLinkValueProps: Set[SmartIri] = allPropertyIris.filter(prop =>
-      allSubPropertyOfRelations(prop).contains(OntologyConstants.KnoraBase.HasLinkToValue.toSmartIri))
-
-    // A set of all subproperties of knora-base:hasFileValue.
-    val allFileValueProps: Set[SmartIri] = allPropertyIris.filter(prop =>
-      allSubPropertyOfRelations(prop).contains(OntologyConstants.KnoraBase.HasFileValue.toSmartIri))
-
-    // A map of the cardinalities defined directly on each resource class. Each class IRI points to a map of
-    // property IRIs to KnoraCardinalityInfo objects.
-    val directClassCardinalities: Map[SmartIri, Map[SmartIri, KnoraCardinalityInfo]] = allClassDefs.map {
-      case (classIri, classDef) =>
-        classIri -> classDef.directCardinalities
-    }
-
-    // Allow each class to inherit cardinalities from its base classes.
-    val classCardinalitiesWithInheritance: Map[SmartIri, Map[SmartIri, KnoraCardinalityInfo]] = allClassIris.map {
-      resourceClassIri =>
-        val resourceClassCardinalities: Map[SmartIri, KnoraCardinalityInfo] = inheritCardinalitiesInLoadedClass(
-          classIri = resourceClassIri,
-          directSubClassOfRelations = directSubClassOfRelations,
-          allSubPropertyOfRelations = allSubPropertyOfRelations,
-          directClassCardinalities = directClassCardinalities
-        )
-
-        resourceClassIri -> resourceClassCardinalities
-    }.toMap
-
-    // Construct a ReadClassInfoV2 for each class.
-    val readClassInfos: Map[SmartIri, ReadClassInfoV2] = makeReadClassInfos(
-      classDefs = allClassDefs,
-      directClassCardinalities = directClassCardinalities,
-      classCardinalitiesWithInheritance = classCardinalitiesWithInheritance,
-      directSubClassOfRelations = directSubClassOfRelations,
-      allSubClassOfRelations = allSubClassOfRelations,
-      allSubPropertyOfRelations = allSubPropertyOfRelations,
-      allPropertyDefs = allPropertyDefs,
-      allKnoraResourceProps = allKnoraResourceProps,
-      allLinkProps = allLinkProps,
-      allLinkValueProps = allLinkValueProps,
-      allFileValueProps = allFileValueProps
-    )
-
-    // Construct a ReadPropertyInfoV2 for each property definition.
-    val readPropertyInfos: Map[SmartIri, ReadPropertyInfoV2] = makeReadPropertyInfos(
-      propertyDefs = allPropertyDefs,
-      directSubPropertyOfRelations = directSubPropertyOfRelations,
-      allSubPropertyOfRelations = allSubPropertyOfRelations,
-      allSubClassOfRelations = allSubClassOfRelations,
-      allGuiAttributeDefinitions = allGuiAttributeDefinitions,
-      allKnoraResourceProps = allKnoraResourceProps,
-      allLinkProps = allLinkProps,
-      allLinkValueProps = allLinkValueProps,
-      allFileValueProps = allFileValueProps
-    )
-
-    // Construct a ReadIndividualV2 for each OWL named individual.
-    val readIndividualInfos = makeReadIndividualInfos(allIndividuals)
-
-    // A ReadOntologyV2 for each ontology to be cached.
-    val readOntologies: Map[SmartIri, ReadOntologyV2] = allOntologyMetadata.map {
-      case (ontologyIri, ontologyMetadata) =>
-        ontologyIri -> ReadOntologyV2(
-          ontologyMetadata = ontologyMetadata,
-          classes = readClassInfos.filter {
-            case (classIri, _) => classIri.getOntologyFromEntity == ontologyIri
-          },
-          properties = readPropertyInfos.filter {
-            case (propertyIri, _) => propertyIri.getOntologyFromEntity == ontologyIri
-          },
-          individuals = readIndividualInfos.filter {
-            case (individualIri, _) => individualIri.getOntologyFromEntity == ontologyIri
-          },
-          isWholeOntology = true
-        )
-    }
-
-    // A set of the IRIs of all properties used in cardinalities in standoff classes.
-    val propertiesUsedInStandoffCardinalities: Set[SmartIri] = readClassInfos.flatMap {
-      case (_, readClassInfo) =>
-        if (readClassInfo.isStandoffClass) {
-          readClassInfo.allCardinalities.keySet
-        } else {
-          Set.empty[SmartIri]
-        }
-    }.toSet
-
-    // A set of the IRIs of all properties whose subject class constraint is a standoff class.
-    val propertiesWithStandoffTagSubjects: Set[SmartIri] = readPropertyInfos.flatMap {
-      case (propertyIri, readPropertyInfo) =>
-        readPropertyInfo.entityInfoContent.getPredicateIriObject(
-          OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri) match {
-          case Some(subjectClassConstraint: SmartIri) =>
-            readClassInfos.get(subjectClassConstraint) match {
-              case Some(subjectReadClassInfo: ReadClassInfoV2) =>
-                if (subjectReadClassInfo.isStandoffClass) {
-                  Some(propertyIri)
-                } else {
-                  None
-                }
-
-              case None => None
-            }
-
-          case None => None
-        }
-    }.toSet
-
-    // Construct the ontology cache data.
-    val ontologyCacheData: OntologyCacheData = OntologyCacheData(
-      ontologies = new ErrorHandlingMap[SmartIri, ReadOntologyV2](readOntologies, { key =>
-        s"Ontology not found: $key"
-      }),
-      subClassOfRelations = new ErrorHandlingMap[SmartIri, Seq[SmartIri]](allSubClassOfRelations, { key =>
-        s"Class not found: $key"
-      }),
-      superClassOfRelations = new ErrorHandlingMap[SmartIri, Set[SmartIri]](allSuperClassOfRelations, { key =>
-        s"Class not found: $key"
-      }),
-      subPropertyOfRelations = new ErrorHandlingMap[SmartIri, Set[SmartIri]](allSubPropertyOfRelations, { key =>
-        s"Property not found: $key"
-      }),
-      guiAttributeDefinitions =
-        new ErrorHandlingMap[SmartIri, Set[SalsahGuiAttributeDefinition]](allGuiAttributeDefinitions, { key =>
-          s"salsah-gui:Guielement not found: $key"
-        }),
-      standoffProperties = propertiesUsedInStandoffCardinalities ++ propertiesWithStandoffTagSubjects
-    )
-
-    // Check property subject and object class constraints.
-
-    readPropertyInfos.foreach {
-      case (propertyIri, readPropertyInfo) =>
-        val allSuperPropertyIris: Set[SmartIri] = allSubPropertyOfRelations.getOrElse(propertyIri, Set.empty[SmartIri])
-
-        readPropertyInfo.entityInfoContent.getPredicateIriObject(
-          OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri) match {
-          case Some(subjectClassConstraint) =>
-            // Each property's subject class constraint, if provided, must be a subclass of the subject class constraints of all its base properties.
-            checkPropertyConstraint(
-              cacheData = ontologyCacheData,
-              internalPropertyIri = propertyIri,
-              constraintPredicateIri = OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri,
-              constraintValueToBeChecked = subjectClassConstraint,
-              allSuperPropertyIris = allSuperPropertyIris,
-              errorSchema = InternalSchema,
-              errorFun = { msg: String =>
-                throw InconsistentRepositoryDataException(msg)
-              }
-            )
-
-            // If the property is defined in a project-specific ontology, its subject class constraint, if provided, must be a Knora resource or standoff class.
-            if (!propertyIri.isKnoraBuiltInDefinitionIri) {
-              val baseClassesOfSubjectClassConstraint = allSubClassOfRelations(subjectClassConstraint)
-
-              if (!(baseClassesOfSubjectClassConstraint.contains(OntologyConstants.KnoraBase.Resource.toSmartIri) ||
-                    baseClassesOfSubjectClassConstraint.contains(OntologyConstants.KnoraBase.StandoffTag.toSmartIri))) {
-                throw InconsistentRepositoryDataException(
-                  s"Property $propertyIri is defined in a project-specific ontology, but its knora-base:subjectClassConstraint, $subjectClassConstraint, is not a subclass of knora-base:Resource or knora-base:StandoffTag")
-              }
-            }
-
-          case None => ()
-        }
-
-        readPropertyInfo.entityInfoContent.getPredicateIriObject(
-          OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri) match {
-          case Some(objectClassConstraint) =>
-            // Each property's object class constraint, if provided, must be a subclass of the object class constraints of all its base properties.
-            checkPropertyConstraint(
-              cacheData = ontologyCacheData,
-              internalPropertyIri = propertyIri,
-              constraintPredicateIri = OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri,
-              constraintValueToBeChecked = objectClassConstraint,
-              allSuperPropertyIris = allSuperPropertyIris,
-              errorSchema = InternalSchema,
-              errorFun = { msg: String =>
-                throw InconsistentRepositoryDataException(msg)
-              }
-            )
-
-          case None =>
-            // A resource property must have an object class constraint, unless it's knora-base:resourceProperty.
-            if (readPropertyInfo.isResourceProp && propertyIri != OntologyConstants.KnoraBase.ResourceProperty.toSmartIri) {
-              throw InconsistentRepositoryDataException(
-                s"Property $propertyIri has no knora-base:objectClassConstraint")
-            }
-        }
-    }
-
-    // Check references between ontologies.
-    checkReferencesBetweenOntologies(ontologyCacheData)
-
-    // Update the cache.
-    storeCacheData(ontologyCacheData)
-  }
-
-  /**
-    * Checks a reference between an ontology entity and another ontology entity to see if the target
-    * is in a non-shared ontology in another project.
-    *
-    * @param ontologyCacheData the ontology cache data.
-    * @param sourceEntityIri   the entity whose definition contains the reference.
-    * @param targetEntityIri   the entity that's the target of the reference.
-    * @param errorFun          a function that throws an exception with the specified message if the reference is invalid.
-    */
-  private def checkOntologyReferenceInEntity(ontologyCacheData: OntologyCacheData,
-                                             sourceEntityIri: SmartIri,
-                                             targetEntityIri: SmartIri,
-                                             errorFun: String => Nothing): Unit = {
-    if (targetEntityIri.isKnoraDefinitionIri) {
-      val sourceOntologyIri = sourceEntityIri.getOntologyFromEntity
-      val sourceOntologyMetadata = ontologyCacheData.ontologies(sourceOntologyIri).ontologyMetadata
-
-      val targetOntologyIri = targetEntityIri.getOntologyFromEntity
-      val targetOntologyMetadata = ontologyCacheData.ontologies(targetOntologyIri).ontologyMetadata
-
-      if (sourceOntologyMetadata.projectIri != targetOntologyMetadata.projectIri) {
-        if (!(targetOntologyIri.isKnoraBuiltInDefinitionIri || targetOntologyIri.isKnoraSharedDefinitionIri)) {
-          errorFun(
-            s"Entity $sourceEntityIri refers to entity $targetEntityIri, which is in a non-shared ontology that belongs to another project")
-        }
-      }
-    }
-  }
-
-  /**
-    * Checks a property definition to ensure that it doesn't refer to any other non-shared ontologies.
-    *
-    * @param ontologyCacheData the ontology cache data.
-    * @param propertyDef       the property definition.
-    * @param errorFun          a function that throws an exception with the specified message if the property definition is invalid.
-    */
-  private def checkOntologyReferencesInPropertyDef(ontologyCacheData: OntologyCacheData,
-                                                   propertyDef: PropertyInfoContentV2,
-                                                   errorFun: String => Nothing): Unit = {
-    // Ensure that the property isn't a subproperty of any property in a non-shared ontology in another project.
-
-    for (subPropertyOf <- propertyDef.subPropertyOf) {
-      checkOntologyReferenceInEntity(
-        ontologyCacheData = ontologyCacheData,
-        sourceEntityIri = propertyDef.propertyIri,
-        targetEntityIri = subPropertyOf,
-        errorFun = errorFun
-      )
-    }
-
-    // Ensure that the property doesn't have subject or object constraints pointing to a non-shared ontology in another project.
-
-    propertyDef.getPredicateIriObject(OntologyConstants.KnoraBase.SubjectClassConstraint.toSmartIri) match {
-      case Some(subjectClassConstraint) =>
-        checkOntologyReferenceInEntity(
-          ontologyCacheData = ontologyCacheData,
-          sourceEntityIri = propertyDef.propertyIri,
-          targetEntityIri = subjectClassConstraint,
-          errorFun = errorFun
-        )
-
-      case None => ()
-    }
-
-    propertyDef.getPredicateIriObject(OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri) match {
-      case Some(objectClassConstraint) =>
-        checkOntologyReferenceInEntity(
-          ontologyCacheData = ontologyCacheData,
-          sourceEntityIri = propertyDef.propertyIri,
-          targetEntityIri = objectClassConstraint,
-          errorFun = errorFun
-        )
-
-      case None => ()
-    }
-  }
-
-  /**
-    * Checks a class definition to ensure that it doesn't refer to any non-shared ontologies in other projects.
-    *
-    * @param ontologyCacheData the ontology cache data.
-    * @param classDef          the class definition.
-    * @param errorFun          a function that throws an exception with the specified message if the property definition is invalid.
-    */
-  private def checkOntologyReferencesInClassDef(ontologyCacheData: OntologyCacheData,
-                                                classDef: ClassInfoContentV2,
-                                                errorFun: String => Nothing): Unit = {
-    for (subClassOf <- classDef.subClassOf) {
-      checkOntologyReferenceInEntity(
-        ontologyCacheData = ontologyCacheData,
-        sourceEntityIri = classDef.classIri,
-        targetEntityIri = subClassOf,
-        errorFun = errorFun
-      )
-    }
-
-    for (cardinalityPropIri <- classDef.directCardinalities.keys) {
-      checkOntologyReferenceInEntity(
-        ontologyCacheData = ontologyCacheData,
-        sourceEntityIri = classDef.classIri,
-        targetEntityIri = cardinalityPropIri,
-        errorFun = errorFun
-      )
-    }
-  }
-
-  /**
-    * Checks references between ontologies to ensure that they do not refer to non-shared ontologies in other projects.
-    *
-    * @param ontologyCacheData the ontology cache data.
-    */
-  private def checkReferencesBetweenOntologies(ontologyCacheData: OntologyCacheData): Unit = {
-    for (ontology <- ontologyCacheData.ontologies.values) {
-      for (propertyInfo <- ontology.properties.values) {
-        checkOntologyReferencesInPropertyDef(
-          ontologyCacheData = ontologyCacheData,
-          propertyDef = propertyInfo.entityInfoContent,
-          errorFun = { msg: String =>
-            throw InconsistentRepositoryDataException(msg)
-          }
-        )
-      }
-
-      for (classInfo <- ontology.classes.values) {
-        checkOntologyReferencesInClassDef(
-          ontologyCacheData = ontologyCacheData,
-          classDef = classInfo.entityInfoContent,
-          errorFun = { msg: String =>
-            throw InconsistentRepositoryDataException(msg)
-          }
-        )
-      }
-    }
-  }
-
-  /**
-    * Given a list of ontology graphs, finds the IRIs of all subjects whose `rdf:type` is contained in a given set of types.
-    *
-    * @param ontologyGraphs a list of ontology graphs.
-    * @param entityTypes    the types of entities to be found.
-    * @return a map of ontology IRIs to sets of the IRIs of entities with matching types in each ontology.
-    */
-  private def getEntityIrisFromOntologyGraphs(ontologyGraphs: Iterable[OntologyGraph],
-                                              entityTypes: Set[IRI]): Map[SmartIri, Set[SmartIri]] = {
-    val entityTypesAsIriLiterals = entityTypes.map(entityType => IriLiteralV2(entityType))
-
-    ontologyGraphs.map { ontologyGraph =>
-      val entityIrisInGraph: Set[SmartIri] = ontologyGraph.constructResponse.statements.foldLeft(Set.empty[SmartIri]) {
-        case (acc, (subjectIri: IriSubjectV2, subjectStatements: Map[SmartIri, Seq[LiteralV2]])) =>
-          val subjectTypeLiterals: Seq[IriLiteralV2] = subjectStatements
-            .getOrElse(OntologyConstants.Rdf.Type.toSmartIri,
-                       throw InconsistentRepositoryDataException(s"Subject $subjectIri has no rdf:type"))
-            .collect {
-              case iriLiteral: IriLiteralV2 => iriLiteral
-            }
-
-          if (subjectTypeLiterals.exists(entityTypesAsIriLiterals.contains)) {
-            acc + subjectIri.value.toSmartIri
-          } else {
-            acc
-          }
-
-        case (acc, _) => acc
-      }
-
-      ontologyGraph.ontologyIri -> entityIrisInGraph
-    }.toMap
-  }
-
-  /**
-    * Given the triplestore's response to `getAllOntologyMetadata.scala.txt`, constructs a map of ontology IRIs
-    * to ontology metadata for the ontology cache.
-    *
-    * @param allOntologyMetadataResponse the triplestore's response to the SPARQL query `getAllOntologyMetadata.scala.txt`.
-    * @return a map of ontology IRIs to ontology metadata.
-    */
-  private def buildOntologyMetadata(
-      allOntologyMetadataResponse: SparqlSelectResult): Map[SmartIri, OntologyMetadataV2] = {
-    allOntologyMetadataResponse.results.bindings.groupBy(_.rowMap("ontologyGraph")).map {
-      case (ontologyGraph: IRI, rows: Seq[VariableResultsRow]) =>
-        val ontologyIri = rows.head.rowMap("ontologyIri")
-
-        if (ontologyIri != ontologyGraph) {
-          throw InconsistentRepositoryDataException(
-            s"Ontology $ontologyIri must be stored in named graph $ontologyIri, but it is in $ontologyGraph")
-        }
-
-        val ontologySmartIri = ontologyIri.toSmartIri
-
-        if (!ontologySmartIri.isKnoraOntologyIri) {
-          throw InconsistentRepositoryDataException(s"Ontology $ontologySmartIri is not a Knora ontology")
-        }
-
-        val ontologyMetadataMap: Map[IRI, String] = rows.map { row =>
-          val pred =
-            row.rowMap.getOrElse("ontologyPred",
-                                 throw InconsistentRepositoryDataException(s"Empty predicate in ontology $ontologyIri"))
-          val obj = row.rowMap.getOrElse(
-            "ontologyObj",
-            throw InconsistentRepositoryDataException(s"Empty object for predicate $pred in ontology $ontologyIri"))
-          pred -> obj
-        }.toMap
-
-        val projectIri: SmartIri = ontologyMetadataMap
-          .getOrElse(
-            OntologyConstants.KnoraBase.AttachedToProject,
-            throw InconsistentRepositoryDataException(s"Ontology $ontologyIri has no knora-base:attachedToProject"))
-          .toSmartIri
-        val ontologyLabel: String =
-          ontologyMetadataMap.getOrElse(OntologyConstants.Rdfs.Label, ontologySmartIri.getOntologyName)
-        val lastModificationDate: Option[Instant] = ontologyMetadataMap
-          .get(OntologyConstants.KnoraBase.LastModificationDate)
-          .map(
-            instant =>
-              stringFormatter.xsdDateTimeStampToInstant(
-                instant,
-                throw InconsistentRepositoryDataException(s"Invalid UTC instant: $instant")))
-        val ontologyVersion: Option[String] = ontologyMetadataMap.get(OntologyConstants.KnoraBase.OntologyVersion)
-
-        ontologySmartIri -> OntologyMetadataV2(
-          ontologyIri = ontologySmartIri,
-          projectIri = Some(projectIri),
-          label = Some(ontologyLabel),
-          lastModificationDate = lastModificationDate,
-          ontologyVersion = ontologyVersion
-        )
-    }
-  }
-
-  /**
-    * Constructs a map of class IRIs to [[ReadClassInfoV2]] instances, based on class definitions loaded from the
-    * triplestore.
-    *
-    * @param classDefs                         a map of class IRIs to class definitions.
-    * @param directClassCardinalities          a map of the cardinalities defined directly on each class. Each resource class
-    *                                          IRI points to a map of property IRIs to [[KnoraCardinalityInfo]] objects.
-    * @param classCardinalitiesWithInheritance a map of the cardinalities defined directly on each class or inherited from
-    *                                          base classes. Each class IRI points to a map of property IRIs to
-    *                                          [[KnoraCardinalityInfo]] objects.
-    * @param directSubClassOfRelations         a map of class IRIs to their immediate base classes.
-    * @param allSubClassOfRelations            a map of class IRIs to all their base classes.
-    * @param allSubPropertyOfRelations         a map of property IRIs to all their base properties.
-    * @param allPropertyDefs                   a map of property IRIs to property definitions.
-    * @param allKnoraResourceProps             a set of the IRIs of all Knora resource properties.
-    * @param allLinkProps                      a set of the IRIs of all link properties.
-    * @param allLinkValueProps                 a set of the IRIs of link value properties.
-    * @param allFileValueProps                 a set of the IRIs of all file value properties.
-    * @return a map of resource class IRIs to their definitions.
-    */
-  private def makeReadClassInfos(classDefs: Map[SmartIri, ClassInfoContentV2],
-                                 directClassCardinalities: Map[SmartIri, Map[SmartIri, KnoraCardinalityInfo]],
-                                 classCardinalitiesWithInheritance: Map[SmartIri, Map[SmartIri, KnoraCardinalityInfo]],
-                                 directSubClassOfRelations: Map[SmartIri, Set[SmartIri]],
-                                 allSubClassOfRelations: Map[SmartIri, Seq[SmartIri]],
-                                 allSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]],
-                                 allPropertyDefs: Map[SmartIri, PropertyInfoContentV2],
-                                 allKnoraResourceProps: Set[SmartIri],
-                                 allLinkProps: Set[SmartIri],
-                                 allLinkValueProps: Set[SmartIri],
-                                 allFileValueProps: Set[SmartIri]): Map[SmartIri, ReadClassInfoV2] = {
-    classDefs.map {
-      case (classIri, classDef) =>
-        val ontologyIri = classIri.getOntologyFromEntity
-
-        // Get the OWL cardinalities for the class.
-        val allOwlCardinalitiesForClass: Map[SmartIri, KnoraCardinalityInfo] =
-          classCardinalitiesWithInheritance(classIri)
-        val allPropertyIrisForCardinalitiesInClass: Set[SmartIri] = allOwlCardinalitiesForClass.keys.toSet
-
-        // Identify the Knora resource properties, link properties, link value properties, and file value properties in the cardinalities.
-        val knoraResourcePropsInClass = allPropertyIrisForCardinalitiesInClass.filter(allKnoraResourceProps)
-        val linkPropsInClass = allPropertyIrisForCardinalitiesInClass.filter(allLinkProps)
-        val linkValuePropsInClass = allPropertyIrisForCardinalitiesInClass.filter(allLinkValueProps)
-        val fileValuePropsInClass = allPropertyIrisForCardinalitiesInClass.filter(allFileValueProps)
-
-        // Make sure there is a link value property for each link property.
-
-        val missingLinkValueProps = linkPropsInClass.map(_.fromLinkPropToLinkValueProp) -- linkValuePropsInClass
-
-        if (missingLinkValueProps.nonEmpty) {
-          throw InconsistentRepositoryDataException(
-            s"Resource class $classIri has cardinalities for one or more link properties without corresponding link value properties. The missing (or incorrectly defined) property or properties: ${missingLinkValueProps
-              .mkString(", ")}")
-        }
-
-        // Make sure there is a link property for each link value property.
-
-        val missingLinkProps = linkValuePropsInClass.map(_.fromLinkValuePropToLinkProp) -- linkPropsInClass
-
-        if (missingLinkProps.nonEmpty) {
-          throw InconsistentRepositoryDataException(
-            s"Resource class $classIri has cardinalities for one or more link value properties without corresponding link properties. The missing (or incorrectly defined) property or properties: ${missingLinkProps
-              .mkString(", ")}")
-        }
-
-        // Make sure that the cardinality for each link property is the same as the cardinality for the corresponding link value property.
-        for (linkProp <- linkPropsInClass) {
-          val linkValueProp: SmartIri = linkProp.fromLinkPropToLinkValueProp
-          val linkPropCardinality: KnoraCardinalityInfo = allOwlCardinalitiesForClass(linkProp)
-          val linkValuePropCardinality: KnoraCardinalityInfo = allOwlCardinalitiesForClass(linkValueProp)
-
-          if (!linkPropCardinality.equalsWithoutGuiOrder(linkValuePropCardinality)) {
-            throw InconsistentRepositoryDataException(
-              s"In class $classIri, the cardinality for $linkProp is different from the cardinality for $linkValueProp")
-          }
-        }
-
-        // The class's direct cardinalities.
-        val directCardinalities: Map[SmartIri, KnoraCardinalityInfo] = directClassCardinalities(classIri)
-
-        val directCardinalityPropertyIris = directCardinalities.keySet
-        val allBaseClasses: Seq[SmartIri] = allSubClassOfRelations(classIri)
-        val isKnoraResourceClass = allBaseClasses.contains(OntologyConstants.KnoraBase.Resource.toSmartIri)
-        val isStandoffClass = allBaseClasses.contains(OntologyConstants.KnoraBase.StandoffTag.toSmartIri)
-        val isValueClass = !(isKnoraResourceClass || isStandoffClass) && allBaseClasses.contains(
-          OntologyConstants.KnoraBase.Value.toSmartIri)
-
-        // If the class is defined in project-specific ontology, do the following checks.
-        if (!ontologyIri.isKnoraBuiltInDefinitionIri) {
-          // It must be either a resource class or a standoff class, but not both.
-          if (!(isKnoraResourceClass ^ isStandoffClass)) {
-            throw InconsistentRepositoryDataException(
-              s"Class $classIri must be a subclass either of knora-base:Resource or of knora-base:StandoffTag (but not both)")
-          }
-
-          // All its cardinalities must be on properties that are defined.
-          val cardinalitiesOnMissingProps = directCardinalityPropertyIris.filterNot(allPropertyDefs.keySet)
-
-          if (cardinalitiesOnMissingProps.nonEmpty) {
-            throw InconsistentRepositoryDataException(
-              s"Class $classIri has one or more cardinalities on undefined properties: ${cardinalitiesOnMissingProps
-                .mkString(", ")}")
-          }
-
-          // It cannot have cardinalities both on property P and on a subproperty of P.
-
-          val maybePropertyAndSubproperty: Option[(SmartIri, SmartIri)] = findPropertyAndSubproperty(
-            propertyIris = allPropertyIrisForCardinalitiesInClass,
-            subPropertyOfRelations = allSubPropertyOfRelations
-          )
-
-          maybePropertyAndSubproperty match {
-            case Some((basePropertyIri, propertyIri)) =>
-              throw InconsistentRepositoryDataException(
-                s"Class $classIri has a cardinality on property $basePropertyIri and on its subproperty $propertyIri")
-
-            case None => ()
-          }
-
-          if (isKnoraResourceClass) {
-            // If it's a resource class, all its directly defined cardinalities must be on Knora resource properties, not including knora-base:resourceProperty or knora-base:hasValue.
-
-            val cardinalitiesOnInvalidProps = directCardinalityPropertyIris.filterNot(allKnoraResourceProps)
-
-            if (cardinalitiesOnInvalidProps.nonEmpty) {
-              throw InconsistentRepositoryDataException(
-                s"Resource class $classIri has one or more cardinalities on properties that are not Knora resource properties: ${cardinalitiesOnInvalidProps
-                  .mkString(", ")}")
-            }
-
-            Set(OntologyConstants.KnoraBase.ResourceProperty, OntologyConstants.KnoraBase.HasValue).foreach {
-              invalidProp =>
-                if (directCardinalityPropertyIris.contains(invalidProp.toSmartIri)) {
-                  throw InconsistentRepositoryDataException(
-                    s"Class $classIri has a cardinality on property $invalidProp, which is not allowed")
-                }
-            }
-
-            // Check for invalid cardinalities on boolean properties.
-            checkForInvalidBooleanCardinalities(
-              classIri = classIri,
-              directCardinalities = directCardinalities,
-              allPropertyDefs = allPropertyDefs,
-              schemaForErrors = InternalSchema,
-              errorFun = { msg: String =>
-                throw InconsistentRepositoryDataException(msg)
-              }
-            )
-
-            // All its base classes with Knora IRIs must also be resource classes.
-            for (baseClass <- classDef.subClassOf) {
-              if (baseClass.isKnoraDefinitionIri && !allSubClassOfRelations(baseClass).contains(
-                    OntologyConstants.KnoraBase.Resource.toSmartIri)) {
-                throw InconsistentRepositoryDataException(
-                  s"Class $classIri is a subclass of knora-base:Resource, but its base class $baseClass is not")
-              }
-            }
-
-            // It must have an rdfs:label.
-            if (!classDef.predicates.contains(OntologyConstants.Rdfs.Label.toSmartIri)) {
-              throw InconsistentRepositoryDataException(s"Class $classIri has no rdfs:label")
-            }
-          } else {
-            // If it's a standoff class, none of its cardinalities must be on Knora resource properties.
-
-            val cardinalitiesOnInvalidProps = directCardinalityPropertyIris.filter(allKnoraResourceProps)
-
-            if (cardinalitiesOnInvalidProps.nonEmpty) {
-              throw InconsistentRepositoryDataException(
-                s"Standoff class $classIri has one or more cardinalities on properties that are Knora resource properties: ${cardinalitiesOnInvalidProps
-                  .mkString(", ")}")
-            }
-
-            // All its base classes with Knora IRIs must also be standoff classes.
-            for (baseClass <- classDef.subClassOf) {
-              if (baseClass.isKnoraDefinitionIri) {
-                if (isStandoffClass && !allSubClassOfRelations(baseClass).contains(
-                      OntologyConstants.KnoraBase.StandoffTag.toSmartIri)) {
-                  throw InconsistentRepositoryDataException(
-                    s"Class $classIri is a subclass of knora-base:StandoffTag, but its base class $baseClass is not")
-                }
-              }
-            }
-          }
-        }
-
-        // Each class must be a subclass of all the classes that are subject class constraints of the properties in its cardinalities.
-        checkSubjectClassConstraintsViaCardinalities(
-          internalClassDef = classDef,
-          allBaseClassIris = allBaseClasses.toSet,
-          allClassCardinalityKnoraPropertyDefs =
-            allPropertyDefs.view.filterKeys(allOwlCardinalitiesForClass.keySet).toMap,
-          errorSchema = InternalSchema,
-          errorFun = { msg: String =>
-            throw InconsistentRepositoryDataException(msg)
-          }
-        )
-
-        val inheritedCardinalities: Map[SmartIri, KnoraCardinalityInfo] = allOwlCardinalitiesForClass
-          .filterNot {
-            case (propertyIri, _) => directCardinalityPropertyIris.contains(propertyIri)
-          }
-
-        // Get the class's standoff data type, if any. A standoff class that has a datatype is a subclass of one of the classes
-        // in StandoffDataTypeClasses.
-
-        val standoffDataType: Set[SmartIri] = allSubClassOfRelations(classIri).toSet
-          .intersect(StandoffDataTypeClasses.getStandoffClassIris.map(_.toSmartIri))
-
-        if (standoffDataType.size > 1) {
-          throw InconsistentRepositoryDataException(
-            s"Class $classIri is a subclass of more than one standoff datatype: ${standoffDataType.mkString(", ")}")
-        }
-
-        // A class can be instantiated if it's in a built-in ontology and marked with knora-base:canBeInstantiated, or if it's
-        // a resource class in a project-specific ontology.
-        val canBeInstantiated = if (ontologyIri.isKnoraBuiltInDefinitionIri) {
-          classDef.predicates
-            .get(OntologyConstants.KnoraBase.CanBeInstantiated.toSmartIri)
-            .flatMap(_.objects.headOption) match {
-            case Some(booleanLiteral: BooleanLiteralV2) => booleanLiteral.value
-            case _                                      => false
-          }
-        } else {
-          isKnoraResourceClass
-        }
-
-        val readClassInfo = ReadClassInfoV2(
-          entityInfoContent = classDef,
-          allBaseClasses = allBaseClasses,
-          isResourceClass = isKnoraResourceClass,
-          isStandoffClass = isStandoffClass,
-          isValueClass = isValueClass,
-          canBeInstantiated = canBeInstantiated,
-          inheritedCardinalities = inheritedCardinalities,
-          knoraResourceProperties = knoraResourcePropsInClass,
-          linkProperties = linkPropsInClass,
-          linkValueProperties = linkValuePropsInClass,
-          fileValueProperties = fileValuePropsInClass,
-          standoffDataType = standoffDataType.headOption match {
-            case Some(dataType: SmartIri) =>
-              Some(
-                StandoffDataTypeClasses.lookup(
-                  dataType.toString,
-                  throw InconsistentRepositoryDataException(s"$dataType is not a valid standoff datatype")))
-
-            case None => None
-          }
-        )
-
-        classIri -> readClassInfo
-    }
-  }
-
-  /**
-    * Checks for invalid cardinalities on boolean properties.
-    *
-    * @param classIri the class IRI.
-    * @param directCardinalities the cardinalities directly defined on the class.
-    * @param allPropertyDefs all property definitions.
-    */
-  def checkForInvalidBooleanCardinalities(classIri: SmartIri,
-                                          directCardinalities: Map[SmartIri, KnoraCardinalityInfo],
-                                          allPropertyDefs: Map[SmartIri, PropertyInfoContentV2],
-                                          schemaForErrors: OntologySchema,
-                                          errorFun: String => Nothing): Unit = {
-    // A cardinality on a property with a boolean object must be 1 or 0-1.
-
-    val invalidCardinalitiesOnBooleanProps: Set[SmartIri] = directCardinalities.filter {
-      case (propertyIri, knoraCardinalityInfo) =>
-        val objectClassConstraintIri = OntologyConstants.KnoraBase.ObjectClassConstraint.toSmartIri
-
-        val propertyObjectClassConstraint: SmartIri = allPropertyDefs(propertyIri).requireIriObject(
-          objectClassConstraintIri,
-          errorFun(s"Property ${propertyIri
-            .toOntologySchema(schemaForErrors)} has no ${objectClassConstraintIri.toOntologySchema(schemaForErrors)}")
-        )
-
-        propertyObjectClassConstraint == OntologyConstants.KnoraBase.BooleanValue.toSmartIri &&
-        !(knoraCardinalityInfo.cardinality == Cardinality.MustHaveOne || knoraCardinalityInfo.cardinality == Cardinality.MayHaveOne)
-    }.keySet
-
-    if (invalidCardinalitiesOnBooleanProps.nonEmpty) {
-      errorFun(
-        s"Class ${classIri.toOntologySchema(schemaForErrors).toSparql} has one or more invalid cardinalities on boolean properties: ${invalidCardinalitiesOnBooleanProps
-          .map(_.toOntologySchema(schemaForErrors).toSparql)
-          .mkString(", ")}")
-    }
-  }
-
-  /**
-    * Constructs a map of property IRIs to [[ReadPropertyInfoV2]] instances, based on property definitions loaded from the
-    * triplestore.
-    *
-    * @param propertyDefs                 a map of property IRIs to property definitions.
-    * @param directSubPropertyOfRelations a map of property IRIs to their immediate base properties.
-    * @param allSubPropertyOfRelations    a map of property IRIs to all their base properties.
-    * @param allSubClassOfRelations       a map of class IRIs to all their base classes.
-    * @param allGuiAttributeDefinitions   a map of `Guielement` IRIs to sets of [[SalsahGuiAttributeDefinition]].
-    * @param allKnoraResourceProps        a set of the IRIs of all Knora resource properties.
-    * @param allLinkProps                 a set of the IRIs of all link properties.
-    * @param allLinkValueProps            a set of the IRIs of link value properties.
-    * @param allFileValueProps            a set of the IRIs of all file value properties.
-    * @return a map of property IRIs to [[ReadPropertyInfoV2]] instances.
-    */
-  private def makeReadPropertyInfos(propertyDefs: Map[SmartIri, PropertyInfoContentV2],
-                                    directSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]],
-                                    allSubPropertyOfRelations: Map[SmartIri, Set[SmartIri]],
-                                    allSubClassOfRelations: Map[SmartIri, Seq[SmartIri]],
-                                    allGuiAttributeDefinitions: Map[SmartIri, Set[SalsahGuiAttributeDefinition]],
-                                    allKnoraResourceProps: Set[SmartIri],
-                                    allLinkProps: Set[SmartIri],
-                                    allLinkValueProps: Set[SmartIri],
-                                    allFileValueProps: Set[SmartIri]): Map[SmartIri, ReadPropertyInfoV2] = {
-    propertyDefs.map {
-      case (propertyIri, propertyDef) =>
-        val ontologyIri = propertyIri.getOntologyFromEntity
-
-        validateGuiAttributes(
-          propertyInfoContent = propertyDef,
-          allGuiAttributeDefinitions = allGuiAttributeDefinitions,
-          errorFun = { msg: String =>
-            throw InconsistentRepositoryDataException(msg)
-          }
-        )
-
-        val isResourceProp = allKnoraResourceProps.contains(propertyIri)
-        val isValueProp =
-          allSubPropertyOfRelations(propertyIri).contains(OntologyConstants.KnoraBase.HasValue.toSmartIri)
-        val isLinkProp = allLinkProps.contains(propertyIri)
-        val isLinkValueProp = allLinkValueProps.contains(propertyIri)
-        val isFileValueProp = allFileValueProps.contains(propertyIri)
-
-        // If the property is defined in a project-specific ontology and is a Knora resource property (a subproperty of knora-base:hasValue or knora-base:hasLinkTo), do the following checks.
-        if (!propertyIri.isKnoraBuiltInDefinitionIri && isResourceProp) {
-          // The property must be a subproperty of knora-base:hasValue or knora-base:hasLinkTo, but not both.
-          if (isValueProp && isLinkProp) {
-            throw InconsistentRepositoryDataException(
-              s"Property $propertyIri cannot be a subproperty of both knora-base:hasValue and knora-base:hasLinkTo")
-          }
-
-          // It can't be a subproperty of knora-base:hasFileValue.
-          if (isFileValueProp) {
-            throw InconsistentRepositoryDataException(
-              s"Property $propertyIri cannot be a subproperty of knora-base:hasFileValue")
-          }
-
-          // Each of its base properties that has a Knora IRI must also be a Knora resource property.
-          for (baseProperty <- propertyDef.subPropertyOf) {
-            if (baseProperty.isKnoraDefinitionIri && !allKnoraResourceProps.contains(baseProperty)) {
-              throw InconsistentRepositoryDataException(
-                s"Property $propertyIri is a subproperty of knora-base:hasValue or knora-base:hasLinkTo, but its base property $baseProperty is not")
-            }
-          }
-
-          // It must have an rdfs:label.
-          if (!propertyDef.predicates.contains(OntologyConstants.Rdfs.Label.toSmartIri)) {
-            throw InconsistentRepositoryDataException(s"Property $propertyIri has no rdfs:label")
-          }
-        }
-
-        // A property is editable if it's in a built-in ontology and marked with knora-base:isEditable,
-        // or if it's a resource property in a project-specific ontology.
-        val isEditable = if (ontologyIri.isKnoraBuiltInDefinitionIri) {
-          propertyDef.predicates
-            .get(OntologyConstants.KnoraBase.IsEditable.toSmartIri)
-            .flatMap(_.objects.headOption) match {
-            case Some(booleanLiteral: BooleanLiteralV2) => booleanLiteral.value
-            case _                                      => false
-          }
-        } else {
-          isResourceProp
-        }
-
-        val propertyEntityInfo = ReadPropertyInfoV2(
-          entityInfoContent = propertyDef,
-          isResourceProp = isResourceProp,
-          isEditable = isEditable,
-          isLinkProp = isLinkProp,
-          isLinkValueProp = isLinkValueProp,
-          isFileValueProp = isFileValueProp,
-          isStandoffInternalReferenceProperty = allSubPropertyOfRelations(propertyIri).contains(
-            OntologyConstants.KnoraBase.StandoffTagHasInternalReference.toSmartIri)
-        )
-
-        propertyIri -> propertyEntityInfo
-    }
-  }
-
-  /**
-    * Constructs a map of OWL named individual IRIs to [[ReadIndividualInfoV2]] instances.
-    *
-    * @param individualDefs a map of OWL named individual IRIs to named individuals.
-    * @return a map of individual IRIs to [[ReadIndividualInfoV2]] instances.
-    */
-  private def makeReadIndividualInfos(
-      individualDefs: Map[SmartIri, IndividualInfoContentV2]): Map[SmartIri, ReadIndividualInfoV2] = {
-    individualDefs.map {
-      case (individualIri, individual) =>
-        individualIri -> ReadIndividualInfoV2(individual)
-    }
-  }
-
-  /**
-    * Given all the OWL named individuals available, constructs a map of `salsah-gui:Guielement` individuals to
-    * their GUI attribute definitions.
-    *
-    * @param allIndividuals all the OWL named individuals available.
-    * @return a map of `salsah-gui:Guielement` individuals to their GUI attribute definitions.
-    */
-  private def makeGuiAttributeDefinitions(
-      allIndividuals: Map[SmartIri, IndividualInfoContentV2]): Map[SmartIri, Set[SalsahGuiAttributeDefinition]] = {
-    val guiElementIndividuals: Map[SmartIri, IndividualInfoContentV2] = allIndividuals.filter {
-      case (_, individual) => individual.getRdfType.toString == OntologyConstants.SalsahGui.GuiElementClass
-    }
-
-    guiElementIndividuals.map {
-      case (guiElementIri, guiElementIndividual) =>
-        val attributeDefs: Set[SalsahGuiAttributeDefinition] =
-          guiElementIndividual.predicates.get(OntologyConstants.SalsahGui.GuiAttributeDefinition.toSmartIri) match {
-            case Some(predicateInfo) =>
-              predicateInfo.objects.map {
-                case StringLiteralV2(attributeDefStr, None) =>
-                  stringFormatter.toSalsahGuiAttributeDefinition(
-                    attributeDefStr,
-                    throw InconsistentRepositoryDataException(
-                      s"Invalid salsah-gui:guiAttributeDefinition in $guiElementIri: $attributeDefStr")
-                  )
-
-                case other =>
-                  throw InconsistentRepositoryDataException(
-                    s"Invalid salsah-gui:guiAttributeDefinition in $guiElementIri: $other")
-              }.toSet
-
-            case None => Set.empty[SalsahGuiAttributeDefinition]
-          }
-
-        guiElementIri -> attributeDefs
-    }
-  }
-
-  /**
-    * Validates the GUI attributes of a resource class property.
-    *
-    * @param propertyInfoContent        the property definition.
-    * @param allGuiAttributeDefinitions the GUI attribute definitions for each GUI element.
-    * @param errorFun                   a function that throws an exception. It will be passed the message to be included in the exception.
-    */
-  private def validateGuiAttributes(propertyInfoContent: PropertyInfoContentV2,
-                                    allGuiAttributeDefinitions: Map[SmartIri, Set[SalsahGuiAttributeDefinition]],
-                                    errorFun: String => Nothing): Unit = {
-    val propertyIri = propertyInfoContent.propertyIri
-    val predicates = propertyInfoContent.predicates
-
-    // Find out which salsah-gui:Guielement the property uses, if any.
-    val maybeGuiElementPred: Option[PredicateInfoV2] =
-      predicates.get(OntologyConstants.SalsahGui.GuiElementProp.toSmartIri)
-    val maybeGuiElementIri: Option[SmartIri] = maybeGuiElementPred.map(
-      _.requireIriObject(throw InconsistentRepositoryDataException(
-        s"Property $propertyIri has an invalid object for ${OntologyConstants.SalsahGui.GuiElementProp}")))
-
-    // Get that Guielement's attribute definitions, if any.
-    val guiAttributeDefs: Set[SalsahGuiAttributeDefinition] = maybeGuiElementIri match {
-      case Some(guiElementIri) =>
-        allGuiAttributeDefinitions.getOrElse(
-          guiElementIri,
-          errorFun(s"Property $propertyIri has salsah-gui:guiElement $guiElementIri, which doesn't exist"))
-
-      case None => Set.empty[SalsahGuiAttributeDefinition]
-    }
-
-    // If the property has the predicate salsah-gui:guiAttribute, syntactically validate the objects of that predicate.
-    val guiAttributes: Set[SalsahGuiAttribute] =
-      predicates.get(OntologyConstants.SalsahGui.GuiAttribute.toSmartIri) match {
-        case Some(guiAttributePred) =>
-          val guiElementIri = maybeGuiElementIri.getOrElse(
-            errorFun(s"Property $propertyIri has salsah-gui:guiAttribute, but no salsah-gui:guiElement"))
-
-          if (guiAttributeDefs.isEmpty) {
-            errorFun(
-              s"Property $propertyIri has salsah-gui:guiAttribute, but $guiElementIri has no salsah-gui:guiAttributeDefinition")
-          }
-
-          // Syntactically validate each attribute.
-          guiAttributePred.objects.map {
-            case StringLiteralV2(guiAttributeObj, None) =>
-              stringFormatter.toSalsahGuiAttribute(
-                s = guiAttributeObj,
-                attributeDefs = guiAttributeDefs,
-                errorFun =
-                  errorFun(s"Property $propertyIri contains an invalid salsah-gui:guiAttribute: $guiAttributeObj")
-              )
-
-            case other =>
-              errorFun(s"Property $propertyIri contains an invalid salsah-gui:guiAttribute: $other")
-          }.toSet
-
-        case None => Set.empty[SalsahGuiAttribute]
-      }
-
-    // Check that all required GUI attributes are provided.
-    val requiredAttributeNames = guiAttributeDefs.filter(_.isRequired).map(_.attributeName)
-    val providedAttributeNames = guiAttributes.map(_.attributeName)
-    val missingAttributeNames: Set[String] = requiredAttributeNames -- providedAttributeNames
-
-    if (missingAttributeNames.nonEmpty) {
-      errorFun(
-        s"Property $propertyIri has one or more missing objects of salsah-gui:guiAttribute: ${missingAttributeNames
-          .mkString(", ")}")
-    }
   }
 
   /**
@@ -3398,6 +2240,43 @@ class OntologyResponderV2(responderData: ResponderData) extends Responder(respon
         iri = ONTOLOGY_CACHE_LOCK_IRI,
         task = () =>
           makeTaskFuture(
+            internalClassIri = internalClassIri,
+            internalOntologyIri = internalOntologyIri
+        )
+      )
+    } yield taskResult
+  }
+
+  /**
+    * Removes cardinalities from class.
+    *
+    * @param deleteCardinalitiesFromClassRequest the request to remove cardinalities.
+    * @return a [[ReadOntologyV2]] in the internal schema, containing the new class definition.
+    */
+  private def deleteCardinalitiesFromClass(
+      deleteCardinalitiesFromClassRequest: DeleteCardinalitiesFromClassRequestV2): Future[ReadOntologyV2] = {
+    for {
+      requestingUser <- FastFuture.successful(deleteCardinalitiesFromClassRequest.requestingUser)
+
+      externalClassIri = deleteCardinalitiesFromClassRequest.classInfoContent.classIri
+      externalOntologyIri = externalClassIri.getOntologyFromEntity
+
+      _ <- checkOntologyAndEntityIrisForUpdate(
+        externalOntologyIri = externalOntologyIri,
+        externalEntityIri = externalClassIri,
+        requestingUser = requestingUser
+      )
+
+      internalClassIri = externalClassIri.toOntologySchema(InternalSchema)
+      internalOntologyIri = externalOntologyIri.toOntologySchema(InternalSchema)
+
+      // Do the remaining pre-update checks and the update while holding a global ontology cache lock.
+      taskResult <- IriLocker.runWithIriLock(
+        apiRequestID = deleteCardinalitiesFromClassRequest.apiRequestID,
+        iri = ONTOLOGY_CACHE_LOCK_IRI,
+        task = () =>
+          DeleteCardinalitiesFromClass.deleteCardinalitiesFromClassTaskFuture(
+            deleteCardinalitiesFromClassRequest = deleteCardinalitiesFromClassRequest,
             internalClassIri = internalClassIri,
             internalOntologyIri = internalOntologyIri
         )
